flex/requirements.txt
xmltodict
apache-beam[gcp]
google-cloud-storage